✨ feat(project-dashboard-services): add services for geo scope, top contributors, top partners, and top primary levers with API integration
