+++ research-indicators/src/app/shared/services/get-geo-scope.service.ts
