+++ research-indicators/src/app/shared/services/get-top-primary-levers.service.ts
@@ -1,19 +1,19 @@
 import { Injectable, inject, signal } from '@angular/core';
 import { ApiService } from '@shared/services/api.service';
-import { PROJECT_DASHBOARD_DEFAULT_LIMIT, ProjectDashboardRankedItem } from '@interfaces/project-dashboard.interface';
+import { TopPrimaryLeverItem } from '@interfaces/project-dashboard.interface';
 
 @Injectable()
-export class GetTopPartnersService {
+export class GetTopPrimaryLeversService {
   apiService = inject(ApiService);
 
   contractId = '';
-  limit = PROJECT_DASHBOARD_DEFAULT_LIMIT;
+  limit = 0;
 
-  list = signal<ProjectDashboardRankedItem[]>([]);
+  list = signal<TopPrimaryLeverItem[]>([]);
   loading = signal(false);
   loadError = signal(false);
 
-  main(contractId: string, limit = PROJECT_DASHBOARD_DEFAULT_LIMIT) {
+  main(contractId: string, limit: number) {
     this.contractId = contractId;
     this.limit = limit;
     void this.update();
@@ -28,8 +28,8 @@ export class GetTopPartnersService {
     this.loadError.set(false);
 
     try {
-      const response = await this.apiService.GET_TopPartners(this.contractId, this.limit);
-      const data = response?.data?.top_partners;
+      const response = await this.apiService.GET_TopPrimaryLevers(this.contractId, this.limit);
+      const data = response?.data?.top_primary_levers;
       this.list.set(Array.isArray(data) ? data : []);
     } catch {
       this.list.set([]);
